+++ src/index.js
@@ -1,37 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Resume from './Resume/Resume';
-import App from './App/App';
-import Header from './Header/Header';
-import ProfileHeader from './ProfileHeader/ProfileHeader';
-import MainContent from './MainContent/MainContent';
-import Weather from './Weather/Weather';
-import AdvanceData from './AdvanceData/AdavanceData';
 import * as serviceWorker from './serviceWorker';
 
-// ReactDOM.render(<App />, document.getElementById('root'));
+import MainPage from './MainPage/MainPage';
 
 // ReactDOM.render(<Resume/>, document.getElementById('root')); // This way we can render the element
-const MainPage = () => {
-  const data = {
-    name: "Swapnil Satish Shinde",
-    rollno: "117A1076",
-    imgurl: "https://avatars3.githubusercontent.com/u/37082605?s=460&v=4"
-  };
-  const mainPage = (
-    <div>
-      <Header/>
-      <ProfileHeader {...data} />
-      <MainContent {...data} />
-      <Weather/>
-      <AdvanceData/>
-      <App/>
-    </div>
-  );
-  
-  return mainPage;
-}
 
 ReactDOM.render(<MainPage/>, document.getElementById('root')); // This way we can render the element
